+++ app/src/main/java/ru/teamdroid/recipecraft/data/model/RecipeIngredients.kt
@@ -7,20 +7,20 @@ class RecipeIngredients(
         var id: Int = 0,
         var idRecipe: Int = 0,
         var idIngredient: Int = 0,
-        var amount: Int = 0,
+        var amount: Double = 0.0,
         var title: String = "",
         var time: Long = 0,
         var portion: Int = 0
 ) : Parcelable {
 
-    constructor(parcel: Parcel) : this(parcel.readInt(), parcel.readInt(), parcel.readInt(), parcel.readInt(), parcel.readString()
+    constructor(parcel: Parcel) : this(parcel.readInt(), parcel.readInt(), parcel.readInt(), parcel.readDouble(), parcel.readString()
             ?: "", parcel.readLong(), parcel.readInt())
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeInt(id)
         parcel.writeInt(idRecipe)
         parcel.writeInt(idIngredient)
-        parcel.writeInt(amount)
+        parcel.writeDouble(amount)
         parcel.writeString(title)
         parcel.writeLong(time)
         parcel.writeInt(portion)
